requirements.txt
pillow
google-generativeai
fal_client
rel
websocket-client
requests
